requirements.txt
blockfrost-python==0.4.3